Correct stride types in revcodec

+++ src/template/revencode.c
@@ -4,7 +4,7 @@ static void _t2(rev_fwd_xform, Int, DIMS)(Int* p);
 
 /* reversible forward lifting transform of 4-vector */
 static void
-_t1(rev_fwd_lift, Int)(Int* p, uint s)
+_t1(rev_fwd_lift, Int)(Int* p, ptrdiff_t s)
 {
   Int x, y, z, w;
   x = *p; p += s;
